Move player movement code into Player class
This allows running the same code on the server, for example
for server-side movement checks.
